moved progression bar to its own class
